requirements.txt
psycopg2-binary==2.9.9
py4j==0.10.9.7
pyspark==3.5.1
